overlays: rootfs: change default camera resolution for cheese
By default, Cheese tries to capture photos/videos with the maxmimum 
resolution the sensor can output. 

This results in a large memory allocation request in kernel space which 
can't be fulfilled due to the constrained size of the CMA space.

In order to prevent this allocation issue, we use a more "reasonable" 
image resolution by default.