cleanup.sh: make sure the script won't fail
In case these recipes are used to create a custom image, some of the 
files we remove in the cleanup script might not be present, leading to a 
build failure.

In order to prevent this, we now use `rm -f`, which should never fail.

Additionally, purge the packages during the autoremove.