image: add wys for audio routing
wys is a small daemon designed to automatically change the audio routing 
for voice calls, and restoring the initial config upon call termination.

We use a modified version which switches the card profile to "Voice 
Call" when in a call, and set it back to "HiFi" when the call ends.